Fetch latest version for references, if not specified
As a developer using Confluent's schema registry, I want to always have a remote
reference pointing at the latest remote version. There are cases where the version is unknown,
or where it drifts between environments. By simply using the latest version, a large burden of
managing versions across environments in our build file is lifted.

As a developer using this plugin, I don't want to necessarily specify the remote
schema version when specifying my references. I need a method of communicating this
in my definition(s).

A new method has been added to the `Subject` class, `addReference(name: String, subject:String)`.
When a reference is not specified with a verion, the compatibility and register tasks will
fetch the latest version from the configured schema registry.